# docs/swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Serverless CMF with Full Rest API and RBAC(Role Base Control Access) System.",
        "title": "MayCMF",
        "contact": {
            "name": "eneus"
        },
        "license": {},
        "version": "0.1.0"
    },
    "host": "127.0.0.1:8088",
    "basePath": "/",
    "paths": {
...
        "/api/v1/users": {
            "get": {
                "tags": [
                    "Manage Users"
                ],
                "summary": "Query data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer User Token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Paging index",
                        "name": "current",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Paging Size",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Username (fuzzy query)",
                        "name": "userName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Real name (fuzzy query)",
                        "name": "realName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Role ID (multiple separated by commas)",
                        "name": "roleIDs",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Status (1: Enable 2: Disable)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search result: {list:List data,pagination:{current:Page index,pageSize:Page size,total:Total number}}",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/schema.UserShow"
                            }
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message: Unauthorized}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message: Server Error}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Manage Users"
                ],
                "summary": "Create data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer User Token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "Create data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/schema.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.User"
                        }
                    },
                    "400": {
                        "description": "{error:{code:0,message: Invalid request parameter}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message: Unauthorized}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message: Server Error}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/users/{id}": {
            "get": {
                "tags": [
                    "Manage Users"
                ],
                "summary": "Query specified data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer User Token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Record ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.User"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message: Unauthorized}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "404": {
                        "description": "{error:{code:0,message: Resource does not exist.}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message: Server Error}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Manage Users"
                ],
                "summary": "Update data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer User Token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Record ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/schema.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.User"
                        }
                    },
                    "400": {
                        "description": "{error:{code:0,message: Invalid request parameter}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message: Unauthorized}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message: Server Error}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Manage Users"
                ],
                "summary": "Delete data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer User Token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Record ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{status:OK}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPStatus"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message: Unauthorized}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message: Server Error}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/users/{id}/disable": {
            "patch": {
                "tags": [
                    "Manage Users"
                ],
                "summary": "Disable data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer User Token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Record ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{status:OK}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPStatus"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message: Unauthorized}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message: Server Error}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/users/{id}/enable": {
            "patch": {
                "tags": [
                    "Manage Users"
                ],
                "summary": "Enable data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer User Token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Record ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{status:OK}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPStatus"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message: Unauthorized}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message: Server Error}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "schema.HTTPError": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error item",
                    "type": "object",
                    "$ref": "#/definitions/schema.HTTPErrorItem"
                }
            }
        },
        "schema.HTTPErrorItem": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Error code",
                    "type": "integer"
                },
                "message": {
                    "description": "Error message",
                    "type": "string"
                }
            }
        },
        "schema.HTTPStatus": {
            "type": "object",
            "properties": {
                "status": {
                    "description": "status(OK)",
                    "type": "string"
                }
            }
        },
...
        "schema.Role": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "created_at": {
                    "description": "Creation time",
                    "type": "string"
                },
                "creator": {
                    "description": "Creator",
                    "type": "string"
                },
                "memo": {
                    "description": "Remarks",
                    "type": "string"
                },
                "name": {
                    "description": "Role Name",
                    "type": "string"
                },
                "permissions": {
                    "description": "Permission permission",
                    "type": "object",
                    "$ref": "#/definitions/schema.RolePermissions"
                },
                "record_id": {
                    "description": "Record ID",
                    "type": "string"
                },
                "sequence": {
                    "description": "Sort value",
                    "type": "integer"
                }
            }
        },
...
        "schema.RolePermissions": {
            "type": "array",
            "items": {
                "type": "object",
                "properties": {
                    "actions": {
                        "description": "Action permission list",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "permission_id": {
                        "description": "Permission ID",
                        "type": "string"
                    },
                    "resources": {
                        "description": "Resource permission list",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "schema.User": {
            "type": "object",
            "required": [
                "real_name",
                "status",
                "user_name"
            ],
            "properties": {
                "created_at": {
                    "description": "Creation time",
                    "type": "string"
                },
                "creator": {
                    "description": "Creator",
                    "type": "string"
                },
                "email": {
                    "description": "Email",
                    "type": "string"
                },
                "password": {
                    "description": "Password",
                    "type": "string"
                },
                "phone": {
                    "description": "Phone number",
                    "type": "string"
                },
                "real_name": {
                    "description": "RealName",
                    "type": "string"
                },
                "record_id": {
                    "description": "Record ID",
                    "type": "string"
                },
                "roles": {
                    "description": "Role authorization",
                    "type": "object",
                    "$ref": "#/definitions/schema.UserRoles"
                },
                "status": {
                    "description": "User Status (1: Enable 2: Disable)",
                    "type": "integer"
                },
                "user_name": {
                    "description": "UserName",
                    "type": "string"
                }
            }
        },
...
        "schema.UserRoles": {
            "type": "array",
            "items": {
                "type": "object",
                "properties": {
                    "role_id": {
                        "type": "string"
                    }
                }
            }
        },
        "schema.UserShow": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Creation time",
                    "type": "string"
                },
                "email": {
                    "description": "Email",
                    "type": "string"
                },
                "phone": {
                    "description": "Phone",
                    "type": "string"
                },
                "real_name": {
                    "description": "RealName",
                    "type": "string"
                },
                "record_id": {
                    "description": "Record ID",
                    "type": "string"
                },
                "roles": {
                    "description": "Roles List",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.Role"
                    }
                },
                "status": {
                    "description": "User Status (1: Enable 2: Disable)",
                    "type": "integer"
                },
                "user_name": {
                    "description": "UserName",
                    "type": "string"
                }
            }
        },
...
    }
}